pyproject.toml
[project]
name = "planinf"
version = "0.1.0"
description = ""
authors = [
    {name = "hyndruide",email = "denis.ulric@gmail.com"}
]
requires-python = ">=3.12"
dependencies = [
    "django (>=6.0.2,<7.0.0)",
    "environs (>=14.5.0,<15.0.0)",
    "djangorestframework (>=3.16.1,<4.0.0)",
    "ortools (>=9.15.6755,<10.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "pytest (>=9.0.2,<10.0.0)",
    "pytest-django (>=4.12.0,<5.0.0)",
    "factory-boy (>=3.3.3,<4.0.0)"
]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "planinf.settings"
pythonpath = ["app"]
python_files = ["tests.py", "test_*.py", "*_tests.py"]
